+++ card/src/components/UserDetails/UserDetails.css
