+++ card/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/app/App';
-import { data } from './modules/card-data'
+import App from './components/App/App';
+import { data } from './modules/CardData'
 
 
 ReactDOM.render(
